CORE-18963 some additional kdocs
